+++ Changelog.md
@@ -4,6 +4,7 @@
 - Allow deletion of assignments with no groups (#6880)
 - Add new routes for `update`, `show`, and `index` actions of the Sections API Controller (#6955)
 - Enable the deletion of Grade Entry Forms that have no grades (#6915)
+- Disallow students from uploading .git file and .git folder in their repository (#6963)
 
 ## [v2.4.5]
 - Add workaround for CSP in Safari < 16 (#6947)
@@ -11,7 +12,6 @@
 - Fix bug where grader distribution graph displays graders multiple times (#6950)
 - Fixed bug where TA Summary table failed to display members (#6949)
 - Make display of group member information consistent across submissions and summary table (#6917)
-- Disallow students from uploading .git folder in their repository (#6963)
 
 ## [v2.4.4]
 - Fix websocket connection when logged in with remote auth (#6912)