+++ Changelog.md
@@ -1,12 +1,14 @@
 # Changelog
 
+## [v2.4.6]
+- Disallow students from uploading .git file and .git folder in their repository (#6963)
+
 ## [v2.4.5]
 - Add workaround for CSP in Safari < 16 (#6947)
 - Improve add button labels on autotest settings page (#6948)
 - Fix bug where grader distribution graph displays graders multiple times (#6950)
 - Fixed bug where TA Summary table failed to display members (#6949)
 - Make display of group member information consistent across submissions and summary table (#6917)
-- Disallow students from uploading .git folder in their repository (#6963)
 
 ## [v2.4.4]
 - Fix websocket connection when logged in with remote auth (#6912)